Projects: ability to select multiple tags at once (#17)
* projects.js: enable multiple tag functionality

With this change, when multiple tags are pressed, only those projects that include all the enabled tags are shown.
The "All" button serves as a filter reset button.

* Rename "All" tag to "Reset" and disable by default

The "All" tag is confusing, since the other tags act as filters (hide things), whereas this tag acts as a filter reset (reveals everything). So the tag is renamed from "All" to "Reset" for clarity, and is not highlighted in the beginning.